fix float/int issue

+++ Python/ogusa/demographics.py
@@ -281,10 +281,10 @@ def get_mort(totpers, min_yr, max_yr, graph=False):
     # Calculate implied mortality rates in sub-bins of mort_rates_all.
     mort_rates_mxyr = mort_rates_all[0:max_yr]
     binsize = (max_yr - min_yr + 1) / totpers
-    num_sub_bins = float(100)
+    num_sub_bins = int(100)
     len_subbins = ((np.float64((max_yr - min_yr + 1) * num_sub_bins)) /
                   totpers)
-    mort_rates_sub = np.zeros(num_sub_bins * max_yr, dtype=np.float64)
+    mort_rates_sub = np.zeros(num_sub_bins * max_yr, dtype=float)
     for i in xrange(max_yr):
         mort_rates_sub[i*num_sub_bins:(i+1)*num_sub_bins] = (1 -
             ((1 - np.float64(mort_rates_mxyr[i])) ** (1 / num_sub_bins)))
@@ -973,4 +973,3 @@ def get_pop_objs(E, S, T, min_yr, max_yr, curr_year, GraphDiag=True):
     return (omega_path_S.T, g_n_SS,
         omega_SSfx[-S:] / omega_SSfx[-S:].sum(), 1-mort_rates_S,
         mort_rates_S, g_n_path, imm_rates_mat.T, omega_S_preTP)
-
